+++ src/api/app.py
@@ -25,14 +25,6 @@ app.secret_key = 'my secret key'
 app.config.from_object(__name__)
 app.config.from_envvar('FLASKR_SETTING', silent=True)
 
-# create base
-engine = create_engine("mysql://root:@localhost:3306/vimmanbot",echo=True)
-db_session = scoped_session(sessionmaker(autocommit=False,
-                            autoflush=False,
-                            bind=engine))
-Base = declarative_base()
-Base.query = db_session.query_property()
-
 from models.model import *
 
 @app.before_request
@@ -41,10 +33,6 @@ def before_request():
 
 @app.teardown_request
 def teardown_request(exception):
-    pass
-
-def clear_session():
-    session.clear()
     pass
 
 from questions import questions
@@ -62,62 +50,13 @@ app.register_blueprint(responses.app, url_prefix="/responses")
 from tweets import tweets
 app.register_blueprint(tweets.app, url_prefix="/tweets")
 
+from users import users
+app.register_blueprint(users.app, url_prefix="/users")
 #@app.route('/')
 #@crossdomain(origin='*')
 #def test():
 #    res = {1: 2}
 #    return jsonify(res)
-
-
-
-@app.route('/login', methods=['GET', 'POST'])
-@crossdomain(origin='*')
-def login():
-    """ ログインします
-    """
-    code = 200
-    error = None
-    #logging.debug(request.form)
-
-    if request.method == 'POST':
-        #if request.form['username'] == app.config['USERNAME'] and request.form['password'] == app.config['PASSWORD']:
-        if request.form['username'] == 'test' and request.form['password'] == 'pass':
-            #session['logged_in'] = True
-            set_username(request.form['username'])
-            return redirect('/#/questions')
-            #return jsonify(status_code=code)
-        else:
-            error = 'wrong'
-            return redirect('/#/login')
-            #return jsonify(error=error)
-
-    return jsonify(status_code=code)
-
-def set_username(username):
-    session['username'] = username
-
-def get_username():
-    return session.get('username')
-
-def is_login():
-    return not not get_username()
-
-def user_check():
-    #logging.debug(is_login())
-    if not is_login():
-        logging.debug(is_login())
-        abort(401)
-        #return redirect('/#/login')
-
-@app.route('/logout', methods=['GET'])
-@crossdomain(origin='*')
-def logout():
-    """ ログアウトします
-    """
-    code = 200
-    clear_session()
-    return redirect('/#/login')
-    #return jsonify(status_code=code)
 
 if __name__ == '__main__':
     app.run()
